Proof of Concept: intercept defineProperty()
A lot of contextify issues seem to be related to the fact that we
cannot intercept defineProperty(), see
https://github.com/nodejs/node/issues/6283.

Here is a proof of concept implementation that gets rid of the
CopyProperties() hack in contextify. For simplicty, only getters from the
descriptors are copied. Also, function declarations are not intercepted,
but that should be easy to do.

It'll be a while until I get this cleanly into V8, but I think once the
V8 API allows for intercepting defineProperty() and function declarations,
a lot of contextify issues can be solved.